functions/google-index-update/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
appwrite==5.0.3
cachetools==5.3.3; python_version >= '3.7'
certifi==2024.6.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
google-auth==2.31.0; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pydantic==2.8.0; python_version >= '3.8'
pydantic-core==2.20.0; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
typing-extensions==4.12.2; python_version < '3.13'
urllib3==2.2.2; python_version >= '3.8'
